+++ selfservice/flow/login/handler.go
@@ -425,6 +425,19 @@ func (h *Handler) createBrowserLoginFlow(w http.ResponseWriter, r *http.Request,
 			q.Set("refresh", "true")
 			r.URL.RawQuery = q.Encode()
 		}
+
+		// on OAuth2 flows, we need to use the RequestURL
+		// as the ReturnTo URL.
+		// This is because a user might want to switch between
+		// different flows, such as login to registration and login to recovery.
+		// After completing a complex flow, such as recovery, we want the user
+		// to be redirected back to the original OAuth2 login flow.
+		if hlr.RequestUrl != "" && h.d.Config().OAuth2ProviderOverrideReturnTo(r.Context()) {
+			// replace the return_to query parameter
+			q := r.URL.Query()
+			q.Set("return_to", hlr.RequestUrl)
+			r.URL.RawQuery = q.Encode()
+		}
 	}
 
 	a, sess, err := h.NewLoginFlow(w, r, flow.TypeBrowser)